+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Service class for usage of the country list in other scenarios
+
 ## [1.0.0] - 2019-12-27
 
 ### Added